+++ src/main/java/com/partnerd/PartnerdApplication.java
@@ -2,7 +2,6 @@ package com.partnerd;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.boot.autoconfigure.domain.EntityScan;
 import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
 import org.springframework.data.jpa.repository.config.EnableJpaRepositories;
 import org.springframework.data.mongodb.repository.config.EnableMongoRepositories;
@@ -13,8 +12,8 @@ import org.springframework.scheduling.annotation.EnableAsync;
 @SpringBootApplication
 @EnableJpaAuditing
 @EnableJpaRepositories(basePackages = "com.partnerd.repository")
-@EnableR2dbcRepositories(basePackages = "com.partnerd.r2dbc")
-@EnableMongoRepositories(basePackages = "com.partnerd.repository.chatRoomRepository.mongoRepository")
+@EnableR2dbcRepositories(basePackages = "com.partnerd.r2dbcRepository")
+@EnableMongoRepositories(basePackages = "com.partnerd.mongoRepository")
 @EnableKafka
 @EnableAsync
 public class PartnerdApplication {
