+++ src/main/java/com/partnerd/config/R2dbcConfig.java
@@ -10,14 +10,13 @@ import org.springframework.context.annotation.Configuration;
 import org.springframework.data.r2dbc.config.AbstractR2dbcConfiguration;
 import org.springframework.data.r2dbc.config.EnableR2dbcAuditing;
 import org.springframework.data.r2dbc.repository.config.EnableR2dbcRepositories;
-import org.springframework.r2dbc.core.DatabaseClient;
 
 import static io.r2dbc.spi.ConnectionFactoryOptions.*;
 
 
 @RequiredArgsConstructor
 @EnableR2dbcRepositories(
-        basePackages = "com.partnerd.r2dbc"
+        basePackages = "com.partnerd.r2dbcRepository"
 )
 @EnableR2dbcAuditing
 @Configuration
