+++ src/main/java/com/partnerd/mongoRepository/ChatMessageRepository.java
@@ -1,4 +1,4 @@
-package com.partnerd.repository.chatRoomRepository.mongoRepository;
+package com.partnerd.mongoRepository;
 
 import com.partnerd.domain.chat.ChatMessage;
 import org.springframework.data.mongodb.repository.MongoRepository;
