+++ src/main/java/com/partnerd/r2dbcRepository/ClubR2DBCRepositoryCustom.java
@@ -1,4 +1,4 @@
-package com.partnerd.r2dbc;
+package com.partnerd.r2dbcRepository;
 
 import com.partnerd.web.dto.homeDTO.response.HomeClubDTO;
 import lombok.RequiredArgsConstructor;
