+++ src/main/java/com/partnerd/service/chatService/impl/ChatQueryServiceImpl.java
@@ -2,7 +2,7 @@ package com.partnerd.service.chatService.impl;
 
 import com.partnerd.converter.ChatConverter;
 import com.partnerd.domain.chat.ChatMessage;
-import com.partnerd.repository.chatRoomRepository.mongoRepository.ChatMessageRepository;
+import com.partnerd.mongoRepository.ChatMessageRepository;
 import com.partnerd.service.chatService.ChatQueryService;
 import com.partnerd.web.dto.chatDTO.ChatDTO;
 import lombok.RequiredArgsConstructor;
