+++ src/main/java/com/partnerd/service/homeService/impl/HomeServiceImpl.java
@@ -4,10 +4,10 @@ import com.partnerd.repository.clubRepository.ClubRepository;
 import com.partnerd.repository.collabPostRepository.collabPost.CollabPostRepository;
 import com.partnerd.repository.projectRepository.ProjectRepository;
 import com.partnerd.repository.projectRepository.PromotionProjectRepository;
-import com.partnerd.r2dbc.ClubR2DBCRepositoryCustom;
-import com.partnerd.r2dbc.CollabPostR2DBCRepositoryCustom;
-import com.partnerd.r2dbc.ProjectR2DBCRepositoryCustom;
-import com.partnerd.r2dbc.PromotionProjectR2DBCRepositoryCustom;
+import com.partnerd.r2dbcRepository.ClubR2DBCRepositoryCustom;
+import com.partnerd.r2dbcRepository.CollabPostR2DBCRepositoryCustom;
+import com.partnerd.r2dbcRepository.ProjectR2DBCRepositoryCustom;
+import com.partnerd.r2dbcRepository.PromotionProjectR2DBCRepositoryCustom;
 import com.partnerd.service.homeService.HomeService;
 import com.partnerd.web.dto.homeDTO.response.HomeClubDTO;
 import com.partnerd.web.dto.homeDTO.response.HomeCollabPostDTO;
